backend/agentflow_core/pyproject.toml
[project]
name = "agentflow-core"
version = "0.1.0"
description = "AgentFlow Core - Multi-Agent Workflow Orchestration Engine"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "AgentFlow Team", email = "team@agentflow.dev"}
]
keywords = ["workflow", "orchestration", "agents", "langgraph", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Web Framework
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    
    # Workflow Engine
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    
    # Data Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Database
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.0",
    
    # HTTP Client
    "httpx>=0.24.0",
    
    # Gemini AI
    "google-generativeai>=0.8.0",
    
    # Image Generation (Google Imagen)
    "pillow>=10.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentflow_core"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=agentflow_core --cov-report=term-missing"
